requirements.txt
altair
numpy
pandas
pydeck
streamlit
fasttext
joblib
nltk
python-Levenshtein
pyspellchecker
scikit-learn